+++ src/common/Toolbar.php
@@ -15,6 +15,11 @@ class Toolbar extends BWrapper implements Renderable
 
     protected $elms = [];
 
+    /**
+     * 当前元素
+     *
+     * @var Bar
+     */
     protected $__elm__;
 
     protected $extKey = '';
@@ -147,5 +152,214 @@ class Toolbar extends BWrapper implements Renderable
         }
 
         throw new \UnexpectedValueException('未知调用:' . $name);
+    }
+
+    /**
+     * 以下为代理方法，当前[$__elm__]生效
+     *
+     */
+
+    /**
+     * Undocumented function
+     *
+     * @param boolean $val
+     * @return $this
+     */
+    public function pullRight($val = true)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->pullRight($val);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param boolean $val
+     * @return $this
+     */
+    public function barPullRight($val = true)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->pullRight($val);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param boolean $val
+     * @param array|string $size
+     * @return $this
+     */
+    public function barUseLayer($val, $size = [])
+    {
+        if ($this->__elm__) {
+            $this->__elm__->useLayer($val, $size);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param boolean $val
+     * @param array|string $size
+     * @return $this
+     */
+    public function barLayerSize($size = [])
+    {
+        if ($this->__elm__) {
+            $this->__elm__->useLayer(true, $size);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barLabel($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->label($val);
+        }
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barIcon($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->icon($val);
+        }
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barHref($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->href($val);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barName($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->href($val);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    function barClass($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->class($val);
+        }
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barAttr($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->attr($val);
+        }
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barStyle($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->style($val);
+        }
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barAddClass($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->addClass($val);
+        }
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barAddAttr($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->addAttr($val);
+        }
+        return $this;
+    }
+
+    /**
+     * Undocumented function
+     *
+     * @param string $val
+     * @return $this
+     */
+    public function barAddStyle($val)
+    {
+        if ($this->__elm__) {
+            $this->__elm__->addStyle($val);
+        }
+        return $this;
     }
 }